fix: createGrid working after pressing "New Grid"
The function wasn't working, because of "replaceWith", in DOM it will
be looking the same, but we can't access the container-variable after
replacing it.

Fix is to use replaceChildren to empty the container before populating
it again.

Also the eventListeners wasn't working either after generating a new grid,
because the querySelector was initialized before with the old squares.
To prevent this, the EventListeners was refactored into a function and
called inside the "createGrid"-function to renew them each time a new
grid is generated.